+++ src/world/Server/WorldSession.h
@@ -37,7 +37,7 @@
 #include <string>
 #include "Objects/MovementInfo.h"
 #include "Logging/Logger.hpp"
-#include "CallBack.h"
+#include "Utilities/CallBack.h"
 
 class Player;
 class WorldPacket;
